Modify costs.py
- Modify class GeneralCost
- Modify class CapitalCost
- Modify class Intangible
