+++ pyscnomics/tools/helper.py
@@ -90,248 +90,6 @@ def check_input(target_func, param: np.ndarray | float | int) -> np.ndarray:
     return param_arr
 
 
-def get_cost_adjustment_by_tax(
-    start_year: int,
-    cost: np.ndarray,
-    expense_year: np.ndarray,
-    project_years: np.ndarray,
-    tax_portion: np.ndarray,
-    tax_rate: np.ndarray | float,
-    tax_discount: float,
-) -> np.ndarray:
-    """
-    Adjust cost based on tax scheme over the prescribed time period.
-
-    Parameters
-    ----------
-    start_year : int
-        The start year of the project.
-    cost : np.ndarray
-        An array of costs to be adjusted.
-    expense_year : np.ndarray
-        An array specifying the expense years for each cost element.
-    project_years : np.ndarray
-        An array of project years.
-    tax_portion: np.ndarray
-        The portion of 'cost' that is subject to tax.
-        Must be an array of length equals to the length of 'cost' array.
-    tax_rate: np.ndarray | float
-        The tax rate to apply. Can be a single value or an array.
-    tax_discount: float
-        The tax discount to apply.
-
-    Returns
-    -------
-    cost_adjusted: np.ndarray
-        An array of adjusted costs after applying tax scheme.
-
-    Notes
-    -----
-    The core operations are as follows:
-    (1) Create 'tax_rate_arr' by calling 'check_input()' function. If 'tax_rate'
-        is given as an array, the length of the array must be consistent with
-        the duration of the project. If it is given as a single value, then create
-        an array of length equal to the project duration with all elements set equal
-        to the single value,
-    (2) Identify the index location of 'tax_rate' based on the associated 'expense_year',
-    (3) Calculate the multiplier:
-        mult = 1.0 + tax_portion * tax_rate_arr[tax_id] * (1.0 - tax_discount),
-    (4) Apply cost adjustment = cost * mult.
-    """
-
-    tax_rate_arr = check_input(target_func=project_years, param=tax_rate)
-    tax_rate_id = (expense_year - start_year).astype("int")
-
-    return cost * (
-        1.0 + tax_portion * tax_rate_arr[tax_rate_id] * (1.0 - tax_discount)
-    )
-
-
-def get_cost_adjustment_by_inflation(
-    start_year: int,
-    cost: np.ndarray,
-    expense_year: np.ndarray,
-    project_years: np.ndarray,
-    year_inflation: np.ndarray,
-    inflation_rate: np.ndarray | float,
-) -> np.ndarray:
-    """
-    Adjusts costs for inflation over the project years.
-
-    This function applies inflation adjustments to costs based on the specified years
-    and inflation rates. It calculates a multiplier for each cost value, adjusting it
-    by inflation rates across the period between `start_year` and the `expense_year`.
-
-    Parameters
-    ----------
-    start_year : int
-        The start year of the project.
-    cost : np.ndarray
-        The array of costs to be adjusted for inflation.
-    expense_year : np.ndarray
-        The year(s) when the expenses occur, corresponding to the provided costs.
-    project_years : np.ndarray
-        An array of years representing the project timeline for inflation adjustments.
-    year_inflation : np.ndarray
-        An array of reference year for inflation calculation.
-    inflation_rate : np.ndarray or float
-        The inflation rate(s) for the project period. If a single float is provided,
-        it applies uniformly across all years; otherwise, an array of rates for
-        each year should be given.
-
-    Returns
-    -------
-    np.ndarray
-        An array of costs adjusted for inflation over the specified project years.
-
-    Notes
-    -----
-    The core operations are as follows:
-    (1) Check 'inflation_rate' whether it is provided as an array or as a single value.
-        If it is given as an array, the length of the array must be consistent with the
-        duration of the project. If it is given as a single value, then create an array
-        of length equal to project duration with all elements set equal to the single
-        value.
-    (2) Parameter 'id_start' identify the index location of 'year_inflation' in array
-        'project_years'. The result is then added by unity. This parameter sets up
-        the first index to slice the 'inflation_rate_arr',
-    (3) Parameter 'id_end' configure the index location of 'expense_year' in array
-        'project_years' based on the associated 'year_inflation'. The result is then
-        added by unity. This parameter sets up the second index to slice the 'inflation_rate_arr',
-    (4) Slice 'inflation_rate_arr' according to 'id_start' and 'id_end'. Add the results
-        by unity, then multiple the associated elements.
-    (5) Cost adjustment is undertaken by multiplication: 'cost' * 'mult'.
-    """
-
-    # Create an array of inflation_rate
-    inflation_rate_arr = check_input(target_func=project_years, param=inflation_rate)
-
-    # Specify the start and end indices to slice the inflation rate array
-    id_start = (
-        np.array(
-            [
-                np.argwhere(year_inflation[i] == project_years).ravel()
-                for i, _ in enumerate(year_inflation)
-            ]
-        ).ravel() + 1
-    ).astype(np.int64)
-
-    id_end = (
-        (expense_year - year_inflation) + (year_inflation - start_year) + 1
-    ).astype(np.int64)
-
-    # Multipliers to adjust cost by inflation
-    mult = np.array(
-        [
-            np.prod(1.0 + inflation_rate_arr[id_start[i]:id_end[i]])
-            for i, _ in enumerate(id_start)
-        ]
-    )
-
-    return cost * mult
-
-
-def apply_cost_adjustment(
-    start_year: int,
-    end_year: int,
-    cost: np.ndarray,
-    expense_year: np.ndarray,
-    project_years: np.ndarray,
-    tax_portion: np.ndarray,
-    tax_rate: np.ndarray | float,
-    tax_discount: float,
-    inflation_rate: np.ndarray | float,
-    year_ref: np.ndarray,
-) -> np.ndarray:
-    """
-    Adjusts cost based on inflation and tax scheme over the specified time period.
-
-    Parameters
-    ----------
-    start_year : int
-        The start year of the project.
-    end_year : int
-        The end year of the project.
-    cost : np.ndarray
-        An array of costs to be adjusted.
-    expense_year : np.ndarray
-        An array specifying the expense years for each cost element.
-    project_years : np.ndarray
-        An array of project years.
-    tax_portion: np.ndarray
-        The portion of 'cost' that is subject to tax (VAT or LBT).
-        Must be an array of length equals to the length of 'cost' array.
-    tax_rate: np.ndarray | float
-        The tax rate to apply (VAT or LBT). Can be a single value or an array.
-    tax_discount: float
-        The tax discount to apply (VAT or LBT).
-    inflation_rate: np.ndarray | float
-        The inflation rate to apply. Can be a single value or an array.
-    year_ref : np.ndarray
-        An array of reference year for inflation calculation.
-
-    Returns
-    -------
-    cost_adjusted: np.ndarray
-        An array of adjusted costs after applying inflation and tax scheme.
-
-    Notes
-    -----
-    The core operations are as follows:
-    (1) Apply cost adjustment due tax scheme by calling function get_cost_adjustment_by_tax().
-    (2) Check 'inflation_rate' whether it is provided as an array or as a single value.
-        If it is given as an array, the length of the array must be consistent with the
-        duration of the project. If it is given as a single value, then create an array
-        of length equal to project duration with all elements set equal to the single
-        value.
-    (3) Parameter 'id_start' identify the index location of 'year_ref' in array
-        'project_years'. The result is then added by unity. This parameter sets up
-        the first index to slice the 'inflation_rate_arr',
-    (4) Parameter 'id_end' configure the index location of 'expense_year' in array
-        'project_years' based on the associated 'year_ref'. The result is then added by unity.
-        This parameter sets up the second index to slice the 'inflation_rate_arr',
-    (5) Slice 'inflation_rate_arr' according to 'id_start' and 'id_end'. Add the results
-        by unity, then multiple the associated elements.
-    (6) Cost adjustment is undertaken by multiplication: 'cost_adjusted_by_tax' * 'mult'.
-    """
-
-    # Cost adjustment due to tax
-    cost_adjusted = get_cost_adjustment_by_tax(
-        start_year=start_year,
-        cost=cost,
-        expense_year=expense_year,
-        project_years=project_years,
-        tax_portion=tax_portion,
-        tax_rate=tax_rate,
-        tax_discount=tax_discount,
-    )
-
-    # Create array of inflation_rate
-    inflation_rate_arr = check_input(target_func=project_years, param=inflation_rate)
-
-    # Specify the start and end indices to slice the inflation_rate array
-    id_start = np.array(
-        [
-            np.argwhere(year_ref[i] == project_years).ravel()[0] + 1
-            for i, val in enumerate(year_ref)
-        ]
-    )
-
-    id_end = (
-        (expense_year - year_ref) + (year_ref - start_year) + 1
-    ).astype("int")
-
-    # Multipliers to adjust cost by inflation
-    mult = np.array(
-        [
-            np.prod(1.0 + inflation_rate_arr[id_start[i]:id_end[i]])
-            for i, val in enumerate(id_start)
-        ]
-    )
-
-    return cost_adjusted * mult
-
-
 def apply_inflation(inflation_rate: np.ndarray | float) -> callable:
     """
     Decorator function to apply inflation/escalation to a target function's result.
